[jay.den] Step 1 & 2 코드리뷰 요청합니다 (#22)
* feat : 토큰 발급 api 작성

- [x] 인증 담당하는 컨트롤러 생성
- [x] 인증 및 토큰 발급 담당하는 서비스 생성

* feat : 토큰 인증 api 적용

- [x] 토큰을 인증하고 유저 정보를 가져오는 ArgumentResolver 생성
- [x] WebMvcConfigure 에 resolver 등록
- [x] 토큰을 이용하여 본인 정보 조회

* feat : 예약하기, 예약취소 개선 - 인증 적용

- [x] 예약 생성시 토큰 확인하여 유효한 경우에만 예약 가능하도록
- [x] 예약 삭제시 토큰 확인하여 유효한 경우에만 예약 취소하도록

* refactor : 예외처리 관련 예외 추가

- 예외처리 진행중

* refactor : 인증 및 토큰 생성 부 예외처리 추가

- 인증 및 토큰 생성 시 회원 정보가 없거나 일치하지 않는 경우에 대한 예외처리 추가
- 해당 관련 테스트 추가

* refactor : 유저정보 조회, 예약 생성 취소 관련 예외처리 추가

- 토큰이 없는 경우, 스케줄이나 예약이 존재하지 않는 경우, 예약의 주인이 아닌 유저가 요청을 하는 경우 등에 대한 예외처리 추가

* refactor : 유저정보 조회, 예약 생성 취소 관련 예외처리 추가

- 토큰이 없는 경우, 스케줄이나 예약이 존재하지 않는 경우, 예약의 주인이 아닌 유저가 요청을 하는 경우 등에 대한 예외처리 추가

* refactor : 프로젝트 구조 변경 및 토큰 검증 추가

- ArgumentResolver 에서 token 의 유효성을 검증하도록 추가
- auth 패키지의 일부 구조를 변경함
- ArgumentResolver 에서 JwtTokenProvider 를 계속 생성하고 있는 부분을 개선하기 위해 WebMvcConfiguration 에서 Resolver 등록시 해당 Resolver 를 생성하면서 JwtTokenProvider 를 주입하도록 수정함

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,8 +1,12 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
+import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
+import nextstep.support.AuthorizationException;
 import nextstep.support.DuplicateEntityException;
+import nextstep.support.NotExistEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
@@ -11,20 +15,23 @@ import java.util.List;
 
 @Service
 public class ReservationService {
-    public final ReservationDao reservationDao;
-    public final ThemeDao themeDao;
-    public final ScheduleDao scheduleDao;
+    private final ReservationDao reservationDao;
+    private final ThemeDao themeDao;
+    private final ScheduleDao scheduleDao;
+    private final MemberDao memberDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, MemberDao memberDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, Long memberId) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
-            throw new NullPointerException();
+            throw new NotExistEntityException();
         }
 
         List<Reservation> reservation = reservationDao.findByScheduleId(schedule.getId());
@@ -32,9 +39,13 @@ public class ReservationService {
             throw new DuplicateEntityException();
         }
 
+        Member member = memberDao.findById(memberId);
+        if (member == null) {
+            throw new NotExistEntityException();
+        }
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                member.getName()
         );
 
         return reservationDao.save(newReservation);
@@ -49,12 +60,15 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long memberId) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
-            throw new NullPointerException();
+            throw new NotExistEntityException();
         }
-
+        Member member = memberDao.findById(memberId);
+        if (member == null || reservation.checkWrongOwner(member.getName())) {
+            throw new AuthorizationException();
+        }
         reservationDao.deleteById(id);
     }
 }
